refactor : 마지막 프레임 공을 굴리는 부분 분할
